+++ packages/uplofile/src/native/index.ts
@@ -1,0 +1,11 @@
+export { Root } from "./Root";
+export type { DocumentPickerResponse } from "react-native-document-picker";
+export { Trigger } from "./Trigger";
+export { Preview } from "./Preview";
+export { useUplofile } from "../shared/hook";
+export { isVideoFile, isImageFile, getExtension, acceptsFile } from "../shared/utils";
+export type {
+  ImageUploaderContextValue, ItemActions, RootProps, UploadFileItem,
+  UploadStatus, UploadResult, UplofileRootRef, BeforeUploadResult,
+  BeforeUploadFn, TriggerRenderProps, PreviewRenderProps,
+} from "../shared/types";